// packages/web/src/shared/api/generated/model/getAnalyticsCalendarParams.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Finnn API
 * Backend API for Finnn finance tracking
 * OpenAPI spec version: 0.1.0
 */

export type GetAnalyticsCalendarParams = {
  dateTo?: string;
  dateFrom?: string;
  description?: string;
  accountIds?: string[];
  categoryIds?: string[];
  transactionTypes?: string[];
  userIds?: string[];
  amountTo?: string;
  amountFrom?: string;
};